requirements.txt
requests~=2.24.0
python-dateutil~=2.8.1
clint~=0.5.1
prettytable~=2.1.0
pytz~=2021.1